+++ sync-for-commerce/src/sdk/models/shared/syncrange.ts
@@ -5,7 +5,7 @@
 import { SpeakeasyBase, SpeakeasyMetadata } from "../../../internal/utils";
 import { Expose, Type } from "class-transformer";
 
-export class SyncRangeDateRange extends SpeakeasyBase {
+export class DateRange extends SpeakeasyBase {
     /**
      * In Codat's data model, dates and times are represented using the <a class="external" href="https://en.wikipedia.org/wiki/ISO_8601" target="_blank">ISO 8601 standard</a>. Date and time fields are formatted as strings; for example:
      *
@@ -64,6 +64,6 @@ export class SyncRangeDateRange extends SpeakeasyBase {
 export class SyncRange extends SpeakeasyBase {
     @SpeakeasyMetadata()
     @Expose({ name: "dateRange" })
-    @Type(() => SyncRangeDateRange)
-    dateRange: SyncRangeDateRange;
+    @Type(() => DateRange)
+    dateRange: DateRange;
 }